news_processor/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-logging
requests
pycountry
python-dateutil